Fix version number in Docker builds
For setuptools_scm to generate the correct version number, the
checkout has to be clean, match the latest tag, and not have any
stray uncommitted files.  We were picking up stray files because
the FastAPI image installs some by default in /app, which we were
using as a working directory.  We were also excluding tests from
the Docker image, which made the checkout dirty.  And we weren't
checking out the full repository with tags so setuptools_scm didn't
see the tags.

Fix all of this by fetching the full repository, changing the
working directory, and removing tests/ from .dockerignore.