Begin migrating to server/client arch
Move the FastAPI application into a server directory; allowing us to add
a "client" library package directory off the repo root in the future.

I've chosen to keep the Dockerfile in the repo's right, even though it's
really for the server. I think this will be helpful when installing the
client package dependency during the docker builds.